adding double as an allowed simple type
The Apimatic code generation adds @var double for fields that are
defined as “Precision” in the UI, which would cause the mapper to throw
a class not found exception.  Adding double to the isSimpleType method.

Added a test for this as well, but that required some updates to
namespaces in the tests in order to get tests to run (seemed like they
weren’t updated after the namespaces were added).