Fix mapping objects implementing ArrayAccess (not collections)
Only handle objects as array when they implement both
ArrayAccess and Traversable.

BC break!

----

Originally, JsonMapper handled objects extending ArrayObject as arrays.

Extending own collection classes from ArrayObject is not always feasible
(issue #175, https://github.com/cweiske/jsonmapper/issues/175),
so a way was sought to rely on interfaces only.

Patch #197 (https://github.com/cweiske/jsonmapper/pull/197) changed
the implementation to check for the ArrayAccess interface instead of
ArrayObject.

This unfortunately breaks
objects-that-allow-array-access-but-are-not-traversable-arrays
(issue #224, https://github.com/cweiske/jsonmapper/pull/224),
for example when you allow array access to properties stored
in some internal variable.

The correct solution is to check that the object implements
ArrayAcces *and* Traversable - then we can be sure the object
is intended to be used with e.g. foreach().

Resolves: https://github.com/cweiske/jsonmapper/pull/224