Condense assertions
Use assertSame to validate both type and value in one step.
Use assertCount to check number of array elements.
Use assertContainsOnly(InstanceOf) to check type of array elements.

Tiny cleanup for tests that are only run for PHP 7.0 or 7.1 and up,
as they need no check for the PHP version anymore.

Call assertions via "$this" everywhere.